Implement callbacks inside useStore and remove Store component (no makes sense now)
