+++ tests/getStore.test.js
@@ -1,0 +1,3 @@
+describe('getStore', () => {
+  it('todo', () => expect(true).toBeTruthy());
+});
