+++ tests/useStore.test.js
@@ -1,0 +1,3 @@
+describe('useStore', () => {
+  it('todo', () => expect(true).toBeTruthy());
+});
