+++ tests/withStore.test.js
@@ -1,0 +1,3 @@
+describe('withStore', () => {
+  it('todo', () => expect(true).toBeTruthy());
+});
